Move all CSS files to seperate styles folder
