+++ client/src/pages/SignOut.tsx
@@ -1,6 +1,5 @@
 // src/pages/SignOut.tsx
 import { Link } from 'react-router-dom';
-import './Home.css';
 
 const Home = () => {
